// ts_gui/src/interfaces/IStoreSettingsJava.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface IStoreSettingsJava { id: string, java17_path: string, java16_path: string, java8_path: string, java_path: string, ram_max_size: number, ram_min_size: number, java_parameter: string, is_built_in_java_vm: boolean, ram_checked: boolean, java_path_checked: boolean, java_parameter_checked: boolean, }